feat: create wishlist blueprint independient of route
