+++ abbybot_website/app/__init__.py
@@ -30,7 +30,9 @@ def create_app():
     # Register Blueprints
     from app.routes.main import main_bp
     from app.routes.abbybot_privileges import abbybot_privileges_bp  # Import the new blueprint
+    from app.routes.wishlist import wishlist_bp  # Import the new wishlist blueprint
     app.register_blueprint(main_bp)
     app.register_blueprint(abbybot_privileges_bp)  # Register the new blueprint
+    app.register_blueprint(wishlist_bp)  # Register the new wishlist blueprint
     
     return app
