+++ abbybot_website/app/templates/emails/wishlist_added.html
@@ -7,11 +7,11 @@
     <title>AbbyBot Wishlist Confirmation</title>
     <style>
         body {
-            background: #c1683a!important;
             color: white !important;
             font-family: 'Inter', sans-serif;
             padding: 20px;
             margin: 0;
+            position: relative;
         }
         .container {
             background-color: #2c2f33;
@@ -21,6 +21,8 @@
             max-width: 800px;
             width: 100%;
             margin: auto;
+            position: relative;
+            z-index: 1;
         }
         .container img {
             max-width: 100%;
@@ -74,14 +76,13 @@
     <table cellpadding="0" cellspacing="0" border="0" width="100%">
         <tr>
             <td align="center">
-                <table cellpadding="0" cellspacing="0" border="0" width="100%" style="max-width: 800px; background-color: #2c2f33; padding: 30px; border-radius: 15px; text-align: center;">
+                <table cellpadding="0" cellspacing="0" border="0" width="100%" class="container">
                     <tr>
                         <td>
                             <h1>Hi {{ username }}, thanks for adding AbbyBot!</h1>
-                            <img src="https://abbybotproject.com/static/img/overview/abbybot_overview.png" alt="AbbyBot - Wishlist Confirmation" style="max-width: 100%; height: auto; width: 150px; margin: 20px 0; border-radius: 10px;">
                             <p>We are thrilled to have you on board. By adding AbbyBot to your wishlist, you will be among the first to know when AbbyBot officially launches.</p>
                             <p>Stay tuned for more updates and exclusive benefits!</p>
-                            <p>As a token of our appreciation, you have obtained the rank <b>Wishlist Users 📜</b>. This rank will be assigned to you when the project enters the beta phase.</p>
+                            <p>As a token of our appreciation, you have obtained the rank <b>Wishlist Users</b>. This rank will be assigned to you when the project enters the beta phase.</p>
                             <p>For more information, <br> <a href="https://abbybotproject.com" class="button" style="background-color: #4caf50; color: white; padding: 15px 25px; text-decoration: none; border-radius: 5px; display: inline-block; margin-top: 20px; font-weight: bold;">AbbyBot Project Website</a></p>
                             <p>Best regards,<br><br>The AbbyBot Project Team &lt;3</p>
                         </td>
