+++ abbybot_website/app/templates/wishlist.html
@@ -28,7 +28,7 @@
                 {% endif %}
             {% endwith %}
             
-            <form action="{{ url_for('main.wishlist') }}" method="POST">
+            <form action="{{ url_for('wishlist.wishlist') }}" method="POST">
                 <div class="field">
                     <label class="label">Name (required field)</label>
                     <div class="control">
